atualizado ex05

+++ src/ex05/Exercicio05.java
@@ -9,11 +9,11 @@ public class Exercicio05 {
         System.out.println("Boas-vindas ao Banco Digital!");
         System.out.println("-----------------------------");
 
-        long numAleatorio = Math.round(Math.random() * 100);
+        int numAleatorio = (int) Math.round(Math.random() * 100);
 
         System.out.printf("Sua posição na fila é: %d %n", numAleatorio);
 
-        for (int i = (int) numAleatorio; i>0; i--){
+        for (int i = numAleatorio; i>0; i--){
             System.out.printf("A fila atualizou! Sua posição é: %d %n", i);
             if (i == 1) {
                 System.out.println("Você será o próximo a ser atendido!");
